refactored code...

+++ DbAccess/Managers/IJobManager.cs
@@ -25,20 +25,20 @@ namespace Exico.HF.DbAccess.Managers
         /// </summary>
         /// <param name="id">Task id of the user task , not HF job id</param>
         /// <returns></returns>
-        Task<bool> Cancel(int id);
+        Task<bool> Cancel(long id);
 
         /// <summary>
         /// Stops the current execution and deletes the job/task as well.
         /// </summary>
         /// <param name="id">Task id of the user task , not HF job id</param>
         /// <returns></returns>
-        Task<bool> Delete(int id);
+        Task<bool> Delete(long id);
 
         /// <summary>
         /// Run a task manually
         /// </summary>
         /// <param name="id">Task id of the user task , not HF job id</param>
         /// <returns></returns>
-        Task RunNow(int id);
+        Task RunNow(long id);
     }
 }
